Refactor Pipeline module DRMAA abstraction into Cluster.py sub-module.
Refactor scripts/farm.py to use abstracted DRMAA functions.

Depends on addition of cluster parameters to cgat/CGAT/Experiment.py
