+++ CGATPipelines/Pipeline/Execution.py
@@ -1,4 +1,4 @@
-##########################################################################
+###########################################################################
 #
 #   MRC FGU Computational Genomics Group
 #
@@ -61,6 +61,7 @@ PARAMS = {}
 from CGATPipelines.Pipeline.Utils import getCallerLocals
 from CGATPipelines.Pipeline.Parameters import substituteParameters
 from CGATPipelines.Pipeline.Files import getTempFilename, getTempFile
+from CGATPipelines.Pipeline.Cluster import *
 
 # global drmaa session
 GLOBAL_SESSION = None
@@ -179,26 +180,6 @@ def execute(statement, **kwargs):
 # detect_pipe_error(): propagate error of programs not at the end of a pipe
 # checkpoint(): exit a set of chained commands (via ;) if the previous
 # command failed.
-_exec_prefix = '''detect_pipe_error_helper()
-    {
-    while [ "$#" != 0 ] ; do
-        # there was an error in at least one program of the pipe
-        if [ "$1" != 0 ] ; then return 1 ; fi
-        shift 1
-    done
-    return 0
-    }
-    detect_pipe_error() {
-    detect_pipe_error_helper "${PIPESTATUS[@]}"
-    return $?
-    }
-    checkpoint() {
-        detect_pipe_error;
-        if [ $? != 0 ]; then exit 1; fi;
-    }
-    '''
-
-_exec_suffix = "; detect_pipe_error"
 
 
 def buildStatement(**kwargs):
@@ -251,34 +232,6 @@ def buildStatement(**kwargs):
     return statement
 
 
-def expandStatement(statement, ignore_pipe_errors=False):
-    '''add generic commands before and after statement.
-
-    The prefixes and suffixes added are defined in :data:`exec_prefix`
-    and :data:`exec_suffix`. The main purpose of these prefixs is to
-    provide error detection code to detect errors at early steps in a
-    series of unix commands within a pipe.
-
-    Arguments
-    ---------
-    statement : string
-        Command line statement to expand
-    ignore_pipe_errors : bool
-        If False, do not modify statement.
-
-    Returns
-    -------
-    statement : string
-        The expanded statement.
-
-    '''
-
-    if ignore_pipe_errors:
-        return statement
-    else:
-        return " ".join((_exec_prefix, statement, _exec_suffix))
-
-
 def joinStatements(statements, infile):
     '''join a chain of statements into a single statement.
 
@@ -326,84 +279,42 @@ def joinStatements(statements, infile):
     return result
 
 
-def getStdoutStderr(stdout_path, stderr_path, tries=5):
-    '''get stdout/stderr allowing for same lag.
+def getJobMemory(options=False, PARAMS=False):
+    '''Extract the job memory from an options or
+       or PARAMS dictionaries'''
 
-    Try at most *tries* times. If unsuccessfull, throw OSError
+    job_memory = None
+    if options and 'job_memory' in options:
+        job_memory = options['job_memory']
 
-    Removes the files once they are read.
+    elif options and "mem_free" in options["cluster_options"] and \
+        PARAMS.get("cluster_memory_resource", False):
 
-    Returns tuple of stdout and stderr.
-    '''
-    x = tries
-    while x >= 0:
-        if os.path.exists(stdout_path):
-            break
-        time.sleep(1)
-        x -= 1
+        E.warn("use of mem_free in job options is deprecated, please"
+               " set job_memory local var instead")
 
-    x = tries
-    while x >= 0:
-        if os.path.exists(stderr_path):
-            break
-        time.sleep(1)
-        x -= 1
+        o = options["cluster_options"]
+        x = re.search("-l\s*mem_free\s*=\s*(\S+)", o)
+        if x is None:
+            raise ValueError(
+                "expecting mem_free in '%s'" % o)
 
-    try:
-        stdout = open(stdout_path, "r").readlines()
-    except IOError, msg:
-        E.warn("could not open stdout: %s" % msg)
-        stdout = []
+        job_memory = x.groups()[0]
 
-    try:
-        stderr = open(stderr_path, "r").readlines()
-    except IOError, msg:
-        E.warn("could not open stdout: %s" % msg)
-        stderr = []
+        # remove memory spec from job options
+        options["cluster_options"] = re.sub(
+            "-l\s*mem_free\s*=\s*(\S+)", "", o)
+    elif PARAMS:
+        job_memory = PARAMS["cluster_memory_default"]
+        # SNS not sure why this was hard coded
+        # PARAMS.get("cluster_memory_default", "2G")
+        # consider better to fail if not set.
+    else:
+        raise ValueError('Job memory must be specified for DRMAA jobs'
+                         ' using either the "job_memory" option or the'
+                         ' "cluster_memory_default" parameter')
 
-    try:
-        os.unlink(stdout_path)
-        os.unlink(stderr_path)
-    except OSError, msg:
-        pass
-
-    return stdout, stderr
-
-
-def _collectSingleJobFromCluster(session, job_id,
-                                 statement,
-                                 stdout_path, stderr_path,
-                                 job_path,
-                                 ignore_errors=False):
-    '''runs a single job on the cluster.'''
-    try:
-        retval = session.wait(
-            job_id, drmaa.Session.TIMEOUT_WAIT_FOREVER)
-    except Exception, msg:
-        # ignore message 24 in PBS code 24: drmaa: Job
-        # finished but resource usage information and/or
-        # termination status could not be provided.":
-        if not msg.message.startswith("code 24"):
-            raise
-        retval = None
-
-    stdout, stderr = getStdoutStderr(stdout_path, stderr_path)
-
-    if retval and retval.exitStatus != 0 and not ignore_errors:
-        raise OSError(
-            "---------------------------------------\n"
-            "Child was terminated by signal %i: \n"
-            "The stderr was: \n%s\n%s\n"
-            "-----------------------------------------" %
-            (retval.exitStatus,
-             "".join(stderr), statement))
-
-    try:
-        os.unlink(job_path)
-    except OSError:
-        E.warn(
-            ("temporary job file %s not present for "
-             "clean-up - ignored") % job_path)
+    return job_memory
 
 
 def run(**kwargs):
@@ -463,123 +374,11 @@ def run(**kwargs):
                                            options['cluster_queue'])
     options['without_cluster'] = options.get('without_cluster')
 
-    job_memory = None
+    # get the memory requirement for the job
+    job_memory = getJobMemory(options, PARAMS)
 
-    if 'job_memory' in options:
-        job_memory = options['job_memory']
-
-    elif "mem_free" in options["cluster_options"] and \
-         PARAMS.get("cluster_memory_resource", False):
-
-        E.warn("use of mem_free in job options is deprecated, please"
-               " set job_memory local var instead")
-
-        o = options["cluster_options"]
-        x = re.search("-l\s*mem_free\s*=\s*(\S+)", o)
-        if x is None:
-            raise ValueError(
-                "expecting mem_free in '%s'" % o)
-
-        job_memory = x.groups()[0]
-
-        # remove memory spec from job options
-        options["cluster_options"] = re.sub(
-            "-l\s*mem_free\s*=\s*(\S+)", "", o)
-    else:
-        job_memory = PARAMS.get("cluster_memory_default", "2G")
-
-    def setupJob(session, options, job_memory, job_name):
-
-        jt = session.createJobTemplate()
-        jt.workingDirectory = PARAMS["workingdir"]
-        jt.jobEnvironment = {'BASH_ENV': '~/.bashrc'}
-        jt.args = []
-        if not re.match("[a-zA-Z]", job_name[0]):
-            job_name = "_" + job_name
-
-        # queue manager specific configuration options
-
-        queue_manager = PARAMS["queue_manager"]
-
-        if queue_manager.lower() == "sge":
-
-            # see: ? cannot find documentation on the SGE native spec
-
-            spec = ["-V",
-                    "-p %(cluster_priority)i",
-                    "-N %s" % job_name,
-                    "%(cluster_options)s"]
-
-            for resource in PARAMS["cluster_memory_resource"].split(","):
-                spec.append("-l %s=%s" % (resource, job_memory))
-
-            # if process has multiple threads, use a parallel environment
-            if 'job_threads' in options:
-                spec.append(
-                    "-pe %(cluster_parallel_environment)s %(job_threads)i -R y")
-            if "cluster_pe_queue" in options and 'job_threads' in options:
-                    spec.append(
-                        "-q %(cluster_pe_queue)s")
-            else:
-                spec.append("-q %(cluster_queue)s")
-
-        elif queue_manager.lower() == "slurm":
-
-            # SLURM DOCS:
-            # http://apps.man.poznan.pl/trac/slurm-drmaa
-            # https://computing.llnl.gov/linux/slurm/cons_res_share.html
-            #
-
-            if 'job_threads' in options:
-                job_threads = options["job_threads"]
-            else:
-                job_threads = 1  # probably should come from a config option
-
-            # Note the that the specified memory must be per CPU
-            # for consistency with the implemented SGE approach
-            if job_memory.endswith("G"):
-                job_memory_per_cpu = int(job_memory[:-1]) * 1000
-            elif job_memory.endswith("M"):
-                job_memory_per_cpu = int(job_memory[:-1])
-            else:
-                raise ValueError('job memory unit not recognised for SLURM, '
-                                 'must be either "M" (for Mb) or "G" (for Gb),'
-                                 ' e.g. 1G or 1000M for 1 Gigabyte of memory')
-
-            # The SLURM Consumable Resource plugin is required
-            # The "CR_CPU_Memory" resource must be specified
-            #
-            # i.e. in slurm.conf:
-            # SelectType=select/cons_res
-            # SelectTypeParameters=CR_CPU_Memory
-            #
-            # * Note that --cpus-per-task will actually refer to cores
-            #   with the appropriate Node configuration
-            #
-            # SLURM-DRMAA DOCS - Note that version 1.2 (SVN) is required
-            # http://apps.man.poznan.pl/trac/slurm-drmaa
-            #
-            # Not implemented:
-            # -V: cannot find documentation as to what this flag is for
-            # -p: does not appear to be part of the slurm drmaa native spec
-            #
-            # TODO: add "--account" (not sure the best way to fill param).
-
-            spec = ["-J %s" % job_name,
-                    "--mem-per-cpu=%s" % job_memory_per_cpu,
-                    "--cpus-per-task=%s" % job_threads,
-                    "%(cluster_options)s"]
-
-        else:
-            raise ValueError("Queue manager %s not supported" % queue_manager)
-
-        jt.nativeSpecification = " ".join(spec) % options
-        # keep stdout and stderr separate
-        jt.joinFiles = False
-
-        E.debug("Job spec is: %s" % jt.nativeSpecification)
-
-        return jt
+    # get the queue manager
+    queue_manager = PARAMS["cluster_queue_manager"]
 
     shellfile = os.path.join(PARAMS["workingdir"], "shell.log")
 
@@ -607,54 +406,37 @@ def run(**kwargs):
         "[:]", "_",
         os.path.basename(options.get("outfile", "ruffus")))
 
-    def buildJobScript(statement, job_memory, job_name):
-        '''build job script from statement.
-
-        returns (name_of_script, stdout_path, stderr_path)
-        '''
-
-        tmpfile = getTempFile(dir=PARAMS["workingdir"])
-        # disabled: -l -O expand_aliases\n" )
-        tmpfile.write("#!/bin/bash\n")
-        tmpfile.write(
-            'echo "%s : START -> %s" >> %s\n' %
-            (job_name, tmpfile.name, shellfile))
+    def _writeJobScript(statement, job_memory, job_name, shellfile):
         # disabled - problems with quoting
         # tmpfile.write( '''echo 'statement=%s' >> %s\n''' %
         # (shellquote(statement), shellfile) )
-        tmpfile.write("set | sed 's/^/%s : /' &>> %s\n" %
-                      (job_name, shellfile))
         # module list outputs to stderr, so merge stderr and stdout
-        tmpfile.write("module list 2>&1 | sed 's/^/%s: /' &>> %s\n" %
-                      (job_name, shellfile))
-        tmpfile.write("hostname | sed 's/^/%s: /' &>> %s\n" %
-                      (job_name, shellfile))
-        tmpfile.write("cat /proc/meminfo | sed 's/^/%s: /' &>> %s\n" %
-                      (job_name, shellfile))
-        tmpfile.write(
-            'echo "%s : END -> %s" >> %s\n' %
-            (job_name, tmpfile.name, shellfile))
+
+        script = '''#!/bin/bash\n
+                    echo "%(job_name)s : START -> ${0}" >> %(shellfile)s
+                    set | sed 's/^/%(job_name)s : /' &>> %(shellfile)s
+                    module list 2>&1 | sed 's/^/%(job_name)s: /' &>> %(shellfile)s
+                    hostname | sed 's/^/%(job_name)s: /' &>> %(shellfile)s
+                    cat /proc/meminfo | sed 's/^/%(job_name)s: /' &>> %(shellfile)s
+                    echo "%(job_name)s : END -> ${0}" >> %(shellfile)s
+                 ''' % locals()
 
         # restrict virtual memory
         # Note that there are resources in SGE which could do this directly
         # such as v_hmem.
         # Note that limiting resident set sizes (RSS) with ulimit is not
         # possible in newer kernels.
-        tmpfile.write("ulimit -v %i\n" % IOTools.human2bytes(job_memory))
+        script += "ulimit -v %i\n" % IOTools.human2bytes(job_memory)
+        script += expandStatement(statement,
+                                  ignore_pipe_errors=ignore_pipe_errors)
+        script += "\n"
 
-        tmpfile.write(
-            expandStatement(
-                statement,
-                ignore_pipe_errors=ignore_pipe_errors) + "\n")
-        tmpfile.close()
+        job_path = getTempFilename(dir=PARAMS["workingdir"])
 
-        job_path = os.path.abspath(tmpfile.name)
-        stdout_path = job_path + ".stdout"
-        stderr_path = job_path + ".stderr"
+        with open(job_path, "w") as script_file:
+            script_file.write(script)
 
-        os.chmod(job_path, stat.S_IRWXG | stat.S_IRWXU)
-
-        return (job_path, stdout_path, stderr_path)
+        return(job_path)
 
     if run_on_cluster:
         # run multiple jobs
@@ -668,29 +450,27 @@ def run(**kwargs):
             if options.get("dryrun", False):
                 return
 
-            jt = setupJob(session, options, job_memory, job_name)
+            jt = setupDrmaaJobTemplate(session, options, job_name, job_memory)
+            E.debug("Job spec is: %s" % jt.nativeSpecification)
 
             job_ids, filenames = [], []
+
             for statement in statement_list:
                 E.debug("running statement:\n%s" % statement)
 
-                job_path, stdout_path, stderr_path = buildJobScript(statement,
-                                                                    job_memory,
-                                                                    job_name)
+                job_path = _writeJobScript(statement, job_memory, job_name, shellfile)
 
-                jt.remoteCommand = job_path
-                jt.outputPath = ":" + stdout_path
-                jt.errorPath = ":" + stderr_path
-
-                os.chmod(job_path, stat.S_IRWXG | stat.S_IRWXU)
+                jt, stdout_path, stderr_path = setDrmaaJobPaths(jt, job_path)
 
                 job_id = session.runJob(jt)
+
                 job_ids.append(job_id)
                 filenames.append((job_path, stdout_path, stderr_path))
 
                 E.debug("job has been submitted with job_id %s" % str(job_id))
 
             E.debug("waiting for %i jobs to finish " % len(job_ids))
+
             session.synchronize(job_ids, drmaa.Session.TIMEOUT_WAIT_FOREVER,
                                 False)
 
@@ -698,7 +478,7 @@ def run(**kwargs):
             for job_id, statement, paths in zip(job_ids, statement_list,
                                                 filenames):
                 job_path, stdout_path, stderr_path = paths
-                _collectSingleJobFromCluster(session, job_id,
+                collectSingleJobFromCluster(session, job_id,
                                              statement,
                                              stdout_path,
                                              stderr_path,
@@ -716,17 +496,11 @@ def run(**kwargs):
             if options.get("dryrun", False):
                 return
 
-            job_path, stdout_path, stderr_path = buildJobScript(statement,
-                                                                job_memory,
-                                                                job_name)
+            jt = setupDrmaaJobTemplate(session, options, job_name, job_memory)
+            E.debug("Job spec is: %s" % jt.nativeSpecification)
 
-            jt = setupJob(session, options, job_memory, job_name)
-
-            jt.remoteCommand = job_path
-            # later: allow redirection of stdout and stderr to files;
-            # can even be across hosts?
-            jt.outputPath = ":" + stdout_path
-            jt.errorPath = ":" + stderr_path
+            job_path = _writeJobScript(statement, job_memory, job_name, shellfile)
+            jt, stdout_path, stderr_path = setDrmaaJobPaths(jt, job_path)
 
             if "job_array" in options and options["job_array"] is not None:
                 # run an array job
@@ -747,7 +521,7 @@ def run(**kwargs):
                 job_id = session.runJob(jt)
                 E.debug("job has been submitted with job_id %s" % str(job_id))
 
-                _collectSingleJobFromCluster(session, job_id,
+                collectSingleJobFromCluster(session, job_id,
                                              statement,
                                              stdout_path,
                                              stderr_path,
